DB Update

+++ database/migrations/2023_03_30_182256_create_payments_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('payments', function (Blueprint $table) {
             $table->id();
             $table->string('payment_for');
-            $table->decimal('amount', 8, 2, true);
+            $table->float('amount', 10, 2, true);
             $table->date('date');
             $table->bigInteger('category_id')->unsigned();
             $table->foreign('category_id')->on('categories')->references('id')->cascadeOnUpdate();
